# specs/openapi.yaml
openapi: 3.0.3
info:
  title: 日程管理系统 API
  version: 3.5.0
  description: |
    v3.3 接口定义 — 前后端契约唯一真相源。

    **v3.3 变更**：
    - 新增任务看板 API：`/tasks`、`/tasks/{id}`、`/tasks/{id}/move`
    - 新增 4 个 schema：TaskProfile、CreateTaskRequest、UpdateTaskRequest、MoveTaskRequest

    **v3.2 变更**：
    - 新增宠物养成 API：`/pets/me`、`/pets/me/interact`、`/shop/items`、`/shop/purchase`
    - 新增 4 个 schema：PetProfile、InteractionResult、ShopItem、PurchaseResult

    **v3.1 变更**：
    - Event 新增 `status` 字段（PLANNED / COMPLETED / CANCELLED），支持"标记完成"。
    - `GET /events` 新增 `tagId`、`status` 过滤参数。
    - 已完成/已取消的日程不再触发提醒，也不参与时间冲突检测。

    **响应约定**：
    - 成功响应直接返回业务数据：单条返回对象，列表返回数组。
    - HTTP 状态码承载语义（200 / 201 / 204 / 400 / 401 / 403 / 404 / 409 / 500）。
    - 异常响应统一为 `ApiResponse`：`{ code, message }`。

    **认证（v3.0）**：
    - 除 `/auth/register` / `/auth/login` / `/auth/refresh` / `/sse/notifications` 外，
      所有端点要求 `Authorization: Bearer <accessToken>`。
    - `/sse/notifications` 不走 Bearer（EventSource 不支持自定义 header），
      改用登录时下发的 `dsa_sse_session` HttpOnly Cookie 鉴权。
    - 用户数据相互隔离：所有 event / category / tag 操作以当前用户为范围。

    变更历史见 `specs/CHANGELOG.md`。
servers:
  - url: /api/v1
security:
  - bearerAuth: []
paths:
...
  /auth/refresh:
    post:
      operationId: refreshToken
      tags: [Auth]
      summary: 使用 refresh token 换取新 access token
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshRequest'
      responses:
        '200':
          description: 续签成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /auth/me:
    get:
      operationId: currentUser
      tags: [Auth]
      summary: 获取当前登录用户信息
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /events:
    get:
      operationId: listEvents
      tags: [Event]
      summary: 按时间范围查询日程
      parameters:
        - name: start
          in: query
          required: true
          schema:
            type: string
            format: date-time
        - name: end
          in: query
          required: true
          schema:
            type: string
            format: date-time
        - name: categoryId
          in: query
          required: false
          schema:
            type: integer
            format: int64
        - name: tagId
          in: query
          required: false
          schema:
            type: integer
            format: int64
          description: 按标签过滤（事件含有该标签即命中）
        - name: status
          in: query
          required: false
          schema:
            $ref: '#/components/schemas/EventStatus'
          description: 按状态过滤
        - name: keyword
          in: query
          required: false
          schema:
            type: string
          description: 按标题、描述、地点搜索
        - name: page
          in: query
          required: false
          schema:
            type: integer
            default: 1
        - name: size
          in: query
          required: false
          schema:
            type: integer
            default: 50
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/EventResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
    post:
      operationId: createEvent
      tags: [Event]
      summary: 创建日程
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EventCreateRequest'
      responses:
        '201':
          description: 创建成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EventResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /events/{id}:
    get:
      operationId: getEventById
      tags: [Event]
      summary: 查询单个日程
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EventResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
    put:
      operationId: updateEvent
      tags: [Event]
      summary: 更新日程
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EventUpdateRequest'
      responses:
        '200':
          description: 更新成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EventResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      operationId: deleteEvent
      tags: [Event]
      summary: 删除日程
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: 删除成功
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /categories:
    get:
      operationId: listCategories
      tags: [Category]
      summary: 获取当前用户的所有分类
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CategoryResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
    post:
      operationId: createCategory
      tags: [Category]
      summary: 创建分类
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CategoryCreateRequest'
      responses:
        '201':
          description: 创建成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CategoryResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '409':
          $ref: '#/components/responses/Conflict'
  /categories/{id}:
    put:
      operationId: updateCategory
      tags: [Category]
      summary: 更新分类
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CategoryCreateRequest'
      responses:
        '200':
          description: 更新成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CategoryResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
    delete:
      operationId: deleteCategory
      tags: [Category]
      summary: 删除分类
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: 删除成功
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        登录后将 `LoginResponse.accessToken` 放入 `Authorization: Bearer <token>` 头。
        access token 默认有效期 15 分钟；过期后用 refresh token 调 `/auth/refresh` 续签。
  responses:
    BadRequest:
      description: 请求参数不合法
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiResponse'
    Unauthorized:
      description: 未认证或 token 失效
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiResponse'
    NotFound:
      description: 资源不存在或无权访问
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiResponse'
    Conflict:
      description: 资源冲突（用户名/邮箱/分类名/标签名重复）
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiResponse'
  schemas:
...
    RefreshRequest:
      type: object
      required: [refreshToken]
      properties:
        refreshToken:
          type: string
    LoginResponse:
      type: object
      required: [accessToken, refreshToken, expiresIn, user]
      properties:
        accessToken:
          type: string
        refreshToken:
          type: string
        expiresIn:
          type: integer
          description: access token 有效期（秒）
        user:
          $ref: '#/components/schemas/UserResponse'
    UserResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        username:
          type: string
        email:
          type: string
        displayName:
          type: string
        avatarUrl:
          type: string
        createdAt:
          type: string
          format: date-time
    EventCreateRequest:
      type: object
      required: [title, startTime, endTime]
      properties:
        title:
          type: string
          maxLength: 200
        description:
          type: string
        startTime:
          type: string
          format: date-time
        endTime:
          type: string
          format: date-time
        allDay:
          type: boolean
          default: false
        location:
          type: string
        color:
          type: string
        reminderMinutes:
          type: integer
        status:
          $ref: '#/components/schemas/EventStatus'
        categoryId:
          type: integer
          format: int64
        tagIds:
          type: array
          items:
            type: integer
            format: int64
    EventUpdateRequest:
      allOf:
        - $ref: '#/components/schemas/EventCreateRequest'
    EventResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        title:
          type: string
        description:
          type: string
        startTime:
          type: string
          format: date-time
        endTime:
          type: string
          format: date-time
        allDay:
          type: boolean
        location:
          type: string
        color:
          type: string
        reminderMinutes:
          type: integer
        status:
          $ref: '#/components/schemas/EventStatus'
        categoryId:
          type: integer
          format: int64
        categoryName:
          type: string
        categoryColor:
          type: string
        tags:
          type: array
          items:
            $ref: '#/components/schemas/TagResponse'
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    EventStatus:
      type: string
      description: |
        日程状态。PLANNED=计划中，COMPLETED=已完成，CANCELLED=已取消。
        创建时缺省按 PLANNED 处理（由服务端保障，不在 schema 声明 default，
        避免代码生成器把 default 套用到查询过滤参数上）。
      enum: [PLANNED, COMPLETED, CANCELLED]
    CategoryCreateRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
          maxLength: 50
        color:
          type: string
        description:
          type: string
    CategoryResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        color:
          type: string
        description:
          type: string
...
    TagResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        color:
          type: string
...
    ApiResponse:
      type: object
      description: 错误响应包装
      properties:
        code:
          type: integer
        message:
          type: string